Move sample graph to data folder
